requirements.txt
streamlit
google-generativeai
sentence-transformers
pandas
numpy
datasets
scikit-learn
torch
